Changed the datatype of all PKs and FKs to long (BigInt) and made a new database for this change to work.

+++ RepairShop/Areas/Admin/Pages/DefectiveUnits/Upsert.cshtml.cs
@@ -305,7 +305,7 @@ namespace RepairShop.Areas.Admin.Pages.DefectiveUnits
             };
         }
 
-        private async Task PopulateMaintenanceContracts(int clientId)
+        private async Task PopulateMaintenanceContracts(long clientId)
         {
             var contracts = (await _unitOfWork.MaintenanceContract.GetAllAsy(
                 mc => mc.IsActive == true && mc.ClientId == clientId,
